added admin editing/profile editing/admin user password reset/user delete functions

+++ controllers/user_controller.js
@@ -27,7 +27,7 @@ const updateAvatar = (req, res) => {
     User.findById(req.params.id)
         .then(user => {
             const id = user._id
-            if (user.avatar.publicId) {
+            if (user.avatar && user.avatar.publicId) {
                 cloudinary.api.delete_resources([user.avatar.publicId],
                     (error, result) => {
                         console.log('deleting avatar from cloud')
@@ -49,4 +49,81 @@ const updateAvatar = (req, res) => {
 
 }
 
-module.exports = { getUsers, updateAvatar }
+const updateDetails = (req, res) => {
+
+    const { firstName, lastName, phone, yearGraduated, country, address } = req.body
+    const newDetails = {
+        firstName: firstName,
+        lastName: lastName,
+        phone: phone,
+        yearGraduated: yearGraduated,
+        country: country,
+        address: address
+    }
+    User.findByIdAndUpdate(req.params.id, newDetails, {new: true})
+        .then(response => {
+            console.log(response)
+            res.send(response)
+        })
+        .catch(err => {
+            console.log(err)
+            res.send(err)
+        })
+}
+
+const changePassword = (req, res) => {
+    
+    const { email, oldPassword, newPassword } = req.body
+
+    User.findByUsername(email)
+        .then(user => {
+            user.changePassword(oldPassword, newPassword)
+                .then(response => {
+                    console.log(response)
+                    res.send(response)
+                })
+                .catch(err => {
+                    console.log(err)
+                    res.send(err)
+                })
+        })
+        .catch(err => {
+            console.log(err)
+            res.send(err)
+        })
+}
+
+const deleteUser = (req, res) => {
+    User.findByIdAndDelete(req.params.id)
+        .then(response => {
+            console.log(response)
+            res.send(response)
+        })
+        .catch(err => {
+            console.log(err)
+            res.send(err)
+        })
+}
+
+const resetPassword = (req, res) => {
+    const { email, reset } = req.body
+    console.log('new password', reset)
+    User.findByUsername(email)
+        .then(user => {
+            user.setPassword(reset)
+                .then(response => {
+                    response.save()
+                        .then(response => {
+                            console.log(response)
+                            res.send(response)
+                        })
+                        .catch(err => res.send(err))
+                })
+                .catch(err => {
+                    console.log(er)
+                    res.send(err)
+                })
+        })
+}
+
+module.exports = { getUsers, updateAvatar, updateDetails, changePassword, deleteUser, resetPassword }
